tests: figureのテストを分割
