--- src/features/editor/extensions/nodes/figure/figure.browser.spec.ts
@@ -1,240 +1,0 @@
-import Document from "@tiptap/extension-document";
-import Paragraph from "@tiptap/extension-paragraph";
-import Text from "@tiptap/extension-text";
-import { userEvent } from "@vitest/browser/context";
-import { describe, expect, it } from "vitest";
-import LakeImage from "@/assets/sikotuko.jpeg";
-import { waitSelectionChange } from "@/tests/dom";
-import { copyText, renderTiptapEditor } from "@/tests/editor";
-import { Details } from "../details";
-import { DetailsContent } from "../details/content";
-import { DetailsSummary } from "../details/summary";
-import { Embed } from "../embed";
-import { EmbedPasteHandler } from "../embed/embed-paste-handler";
-import { Figure } from ".";
-import { Caption } from "./caption";
-import { Image } from "./image";
-
-describe("InputRule", () => {
-  it("![alt](src) で Figure ノードが作成される", async () => {
-    const editor = renderTiptapEditor({
-      content: "<p></p>",
-      extensions: [Document, Paragraph, Text, Figure, Caption, Image],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().run();
-    });
-    await userEvent.keyboard(`!{\\[}支笏湖{\\]}(${LakeImage}) `);
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe("doc(figure(image, caption))");
-    expect(editor.state.selection.from).toBe(3);
-    expect(editor.state.doc.firstChild?.firstChild?.attrs).toEqual({
-      src: LakeImage,
-      alt: "支笏湖",
-      width: null,
-    });
-  });
-
-  it("行の途中では InputRule が発動しない", async () => {
-    const editor = renderTiptapEditor({
-      content: "<p>Text</p>",
-      extensions: [Document, Paragraph, Text, Figure, Caption, Image],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(2).run();
-    });
-    await userEvent.keyboard(`!{\\[}支笏湖{\\]}(${LakeImage}) `);
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(`doc(paragraph("T![支笏湖](${LakeImage}) ext"))`);
-  });
-
-  it("アコーディオンのタイトルでは InputRule が発動しない", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details><summary></summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Figure,
-        Caption,
-        Image,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(2).run();
-    });
-    await userEvent.keyboard(`!{\\[}支笏湖{\\]}(${LakeImage}) `);
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(
-      `doc(details(detailsSummary("![支笏湖](${LakeImage}) "), detailsContent(paragraph("Text"))))`,
-    );
-  });
-});
-
-describe("キー入力", () => {
-  it("キャプションの先頭で Backspace を押すとFigureが削除される", async () => {
-    const editor = renderTiptapEditor({
-      content: `<p><img src='${LakeImage}' alt='支笏湖'><em>支笏湖</em></p>`,
-      extensions: [Document, Paragraph, Text, Image, Caption, Figure],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(3).run();
-    });
-    await userEvent.keyboard("{Backspace}");
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe("doc(paragraph)");
-    expect(editor.state.selection.from).toBe(1);
-  });
-
-  it("キャプション内で Enter を押すと、段落を挿入して移動する", async () => {
-    const editor = renderTiptapEditor({
-      content: `<p><img src='${LakeImage}' alt='支笏湖'><em>支笏湖</em></p>`,
-      extensions: [Document, Paragraph, Text, Image, Caption, Figure],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(4).run();
-    });
-    await userEvent.keyboard("{Enter}");
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(`doc(figure(image, caption("支笏湖")), paragraph)`);
-    expect(editor.state.selection.from).toBe(9);
-  });
-
-  it("キャプションの先頭で左矢印キーを押すと前のノードに移動する", async () => {
-    const editor = renderTiptapEditor({
-      content: `<p>Before</p><p><img src='${LakeImage}' alt='支笏湖'><em>支笏湖</em></p>`,
-      extensions: [Document, Paragraph, Text, Figure, Caption, Image],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(11).run();
-    });
-
-    await userEvent.keyboard("{ArrowLeft}");
-
-    expect(editor.state.selection.from).toBe(7); // "Before" の最後
-  });
-
-  it("キャプションの末尾で右矢印キーを押すと次のノードに移動する", async () => {
-    const editor = renderTiptapEditor({
-      content: `<p><img src='${LakeImage}' alt='支笏湖'><em>支笏湖</em></p><p>After</p>`,
-      extensions: [Document, Paragraph, Text, Figure, Caption, Image],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(6).run();
-    });
-
-    await userEvent.keyboard("{ArrowRight}");
-
-    expect(editor.state.selection.from).toBe(9); // "After" の最初
-  });
-});
-
-describe("ペースト", () => {
-  it("![alt](src) をペーストすると Figure ノードが作成される", async () => {
-    const editor = renderTiptapEditor({
-      content: "",
-      extensions: [Document, Paragraph, Text, Figure, Caption, Image],
-    });
-
-    const input = document.createElement("input");
-    input.value = `![支笏湖](${LakeImage})`;
-    document.body.appendChild(input);
-
-    await userEvent.tripleClick(input);
-    await userEvent.copy();
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().run();
-    });
-
-    await userEvent.paste();
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe("doc(figure(image, caption))");
-    expect(editor.state.selection.from).toBe(3);
-    expect(editor.state.doc.firstChild?.firstChild?.attrs).toEqual({
-      src: LakeImage,
-      alt: "支笏湖",
-      width: null,
-    });
-
-    input.remove(); // クリーンアップ
-  });
-
-  it("拡張子がjpegの画像URLをペーストするとFigureノードが作成される(埋め込みより優先)", async () => {
-    const editor = renderTiptapEditor({
-      content: "",
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Figure,
-        Caption,
-        Image,
-        Embed,
-        EmbedPasteHandler,
-      ],
-    });
-    const url = `${location.origin}${LakeImage}`;
-
-    await copyText(url);
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(1).run();
-    });
-    await userEvent.paste();
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe("doc(figure(image, caption))");
-    expect(editor.state.selection.from).toBe(3);
-    expect(editor.state.doc.firstChild?.firstChild?.attrs).toEqual({
-      src: url,
-      alt: "",
-      width: null,
-    });
-  });
-
-  it("アコーディオンのタイトルでペーストしても、Figureノードが生成されない", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details><summary></summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Figure,
-        Caption,
-        Image,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await copyText(`![支笏湖](${LakeImage})`);
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(2).run();
-    });
-    await userEvent.paste();
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(
-      `doc(details(detailsSummary("![支笏湖](${LakeImage})"), detailsContent(paragraph("Text"))))`,
-    );
-  });
-});
